move files based on updated package structure
OCD-2159
