+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/changerequest/dao/ChangeRequestDAO.java
@@ -1,0 +1,10 @@
+package gov.healthit.chpl.changerequest.dao;
+
+import gov.healthit.chpl.changerequest.domain.ChangeRequest;
+import gov.healthit.chpl.exception.EntityRetrievalException;
+
+public interface ChangeRequestDAO {
+    ChangeRequest create(final ChangeRequest cr) throws EntityRetrievalException;
+
+    ChangeRequest get(final Long changeRequestId) throws EntityRetrievalException;
+}
