+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/changerequest/dao/ChangeRequestStatusTypeDAO.java
@@ -1,0 +1,13 @@
+package gov.healthit.chpl.changerequest.dao;
+
+import java.util.List;
+
+import gov.healthit.chpl.changerequest.domain.ChangeRequestStatusType;
+import gov.healthit.chpl.exception.EntityRetrievalException;
+
+public interface ChangeRequestStatusTypeDAO {
+    ChangeRequestStatusType getChangeRequestStatusTypeById(Long changeRequestStatusTypeId)
+            throws EntityRetrievalException;
+
+    List<ChangeRequestStatusType> getChangeRequestStatusTypes();
+}
