+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/changerequest/dao/ChangeRequestWebsiteDAO.java
@@ -1,0 +1,12 @@
+package gov.healthit.chpl.changerequest.dao;
+
+import gov.healthit.chpl.changerequest.domain.ChangeRequest;
+import gov.healthit.chpl.changerequest.domain.ChangeRequestWebsite;
+import gov.healthit.chpl.exception.EntityRetrievalException;
+
+public interface ChangeRequestWebsiteDAO {
+    ChangeRequestWebsite create(final ChangeRequest cr, final ChangeRequestWebsite crWebsite)
+            throws EntityRetrievalException;
+
+    ChangeRequestWebsite getByChangeRequestId(final Long changeRequestId) throws EntityRetrievalException;
+}
