+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/changerequest/domain/ChangeRequestWebsite.java
@@ -1,0 +1,26 @@
+package gov.healthit.chpl.changerequest.domain;
+
+import java.io.Serializable;
+
+public class ChangeRequestWebsite implements Serializable {
+    private static final long serialVersionUID = -5572794875424284955L;
+
+    private Long id;
+    private String website;
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(final Long id) {
+        this.id = id;
+    }
+
+    public String getWebsite() {
+        return website;
+    }
+
+    public void setWebsite(String website) {
+        this.website = website;
+    }
+}
