--- chpl/chpl-service/src/main/java/gov/healthit/chpl/dao/changerequest/ChangeRequestDAO.java
@@ -1,10 +1,0 @@
-package gov.healthit.chpl.dao.changerequest;
-
-import gov.healthit.chpl.domain.changerequest.ChangeRequest;
-import gov.healthit.chpl.exception.EntityRetrievalException;
-
-public interface ChangeRequestDAO {
-    ChangeRequest create(final ChangeRequest cr) throws EntityRetrievalException;
-
-    ChangeRequest get(final Long changeRequestId) throws EntityRetrievalException;
-}
